+++ src/components/App.jsx
@@ -9,7 +9,6 @@ import {
   CalculatorPage,
 } from 'pages';
 
-
 export const App = () => {
   return (
     <div>
