+++ src/pages/index.js
@@ -1,5 +1,5 @@
 export { default as HomePage } from './HomePage';
-export { default as RegistrationPage } from './RegisterPage';
+export { default as RegistrationPage } from '../components/RegistrationForm/RegistrationForm';
 export { default as LoginPage } from './LoginPage';
 export { default as DiaryPage } from './DiaryPage';
 export { default as CalculatorPage } from './CalculatorPage';
