+++ src/sass/base/BaseStyles.scss
@@ -138,7 +138,7 @@ select:-webkit-autofill:focus {
 }
 
 @mixin orangeBtn {
-    width: 176px;
+    width: 182px;
     background-color: $accent-color;
     color: $accent-text-color;
     padding: 13px 0;
