+++ src/sass/main.scss
@@ -4,5 +4,5 @@
 @import "./components/header";
 @import "../../src/components/DiaryAddProductForm/DiaryAddProductForm.module.scss";
 @import '../../src/components/DiaryProductsList/DiaryProductsList.module.scss';
-@import "../../src/components/DiaryProductsListItem/DiaryProductsListItem.module.scss"
-
+@import "../../src/components/DiaryProductsListItem/DiaryProductsListItem.module.scss";
+@import '../../src/components/RegistrationForm/RegisterForm.module.scss';
